+++ packages/common/enums/tosu.ts
@@ -1,0 +1,4 @@
+export enum ClientType {
+    stable,
+    lazer
+}
